Warning fix
git-svn-id: https://svn.oktetlabs.ru/svnroot/prj/oktetlabs/test-environment/src/trunk@29410 9d14d920-75d5-0310-8bd7-a52572981aee

+++ engine/configurator/conf_ta.c
@@ -430,7 +430,7 @@ sync_ta_subtree(const char *ta, const char *oid)
         return rc;
     }
 
-    if (cfg_get_buf_len < strlen(cfg_get_buf) + strlen(oid) + 2)
+    if (cfg_get_buf_len < (int)strlen(cfg_get_buf) + (int)strlen(oid) + 2)
     {
         cfg_get_buf_len <<= 1;
